trace/preemptirq: Let the tracepoints know why preemption was disabled
Forward the reason why the preemption was disabled to the tracepoints.

It does not print the reason, but the field is there for the functions
that hook the tracer to interpret it.

The IRQ control can also be updated, adding the information if the
current thread masked IRQs, or if it is caused by an annotation of an
kernel entry point, notifying that the hardware disabled the interrupts,
e.g., at the beginning of the interrupt handlers.

No functional changes.

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>