trace/preemptirq: Add helper functions for switching to the scheduling context
For optimization reasons, if the preemption was already disabled to
postpone the scheduler, it will not be re enabled and disabled back to
call the scheduler. It makes sense.

So, we need to add an annotation for this context. These helper
functions do that.

No functional change, and no code add if the preempt tracing is off.

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>