sched/idle: Annotate the preempt disable from idle tasks
Idle runs with preempt disabled. It actually does not disable
preemption: the preemption is not enabled in the return from scheduler,
and so it is not disabled before calling back the scheduler.

Annotate the call and the return of the scheduler by switching the
reason why the preemption was disabled.

No functional changes, and the code is only added if preemption tracer
is enabled.

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>