rcu: Set rcu_irq_[enter|exit]_irqson notrace
rcu_irq_enter_irqson() and rcu_irq_exit_irqson() are used by the tracing
subsystem, mainly when tracing in code entry points.

What is happening is that, the preemptirq:irq_disable/enable
tracepoints, which are called in the interrupt entry, are causing
a recursion with the local_irq_save/restore used in
rcu_irq_exit_irqson() and rcu_irq_enter_irqson().

To avoid this recursion, set these functions as notrace, and use the
notrace versions of raw_local_irq_save/restore.

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>