rtsl: Real-Time Scheduling Latency Tracer
This is the latency parse module presented in the paper:

 Demystifying the Real-Time Linux Scheduling Latency
  by: Daniel Bristot de Oliveira, Daniel Casini, Rômulo Silva de
  Oliveira and Tommaso Cucinotta.

In the paper, the real-time Linux latency is defined using four
variables:

    - DST:  the sched tail delay which is the delay from the IRQs
            being disabled to cause the context switch and the
            return from the scheduler.
    - POID: the longest preemption or IRQ disabled to postpone
            the scheduler;
    - PAIE: the longest time in which the preemption and IRQs are
            transiently enabled in the return of the preemption or
            IRQ enable, that will cause the scheduler execution to
            preempt the current thread.
    - PSD:  the longest time in which the preemption is disabled
            to execute __schedule() function.

plus the interference from interrupts (IRQs and NMI).

This tracer provides a method to trace the value for this
variables, while trying to keep the overhead as low as possible.

The value for these variables is exported via tracepoints. The
reason for that is that tracepoints are flexible, enabling the
trace using perf/ftrace/bpf, and also supports the hook of
other code on top of it.

The tracepoints add by this module are:
	rtsl:dst		a DST occurrence
	rtsl:poid		a POID occurrence
	rtsl:paie		a PAIE occurrence
	rtsl:psd		a PSD occurrence

	rtsl:max_dst		a new max value for DST
	rtsl:max_poid		a new max value for POID
	rtsl:max_paie		a new max value for PAIE
	rtsl:max_psd		a new max value for PSD

	rtsl:nmi_execution	NMI occurence
	rtsl:irq_execution	IRQ occurence

The rtsl:max_* events exist to avoid printing the high
frequency rtsl:dst and rtsl:p* events, by only printing
a new "max value", reducing the overhead.

It is worth mentioning that these are "net" values, that is,
the interference from other variables is discounted. For
example, if an IRQ takes place during a POID, the IRQ
duration is discounted from the POID. If an NMI takes
place during that IRQ, the NMI duration is discounted
from the IRQ and POID, and so on.

These tracepoints work hooking to other tracepoints. By default,
they are not hooked. To enable them, you need to:

and start having fun! This also reset known max values.

to unhook the tracepoints.

This feature depends on PREEMPT_RT and PREEMPTIRQ tracepoints.

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>